added square wave shape with duty cycle
square shape was already possible by setting the shapeCurve to 0 for a sine wave, but now it's also possible to use a square curve directly, and then the shapeCurve is used to control the duty cycle.
This opens up a lot more sound options without new parameters and very little code.